Preparations for x64 support

+++ Example/Examples/src/Advanced/FirstGpgpu/FirstGpgpu.cpp
@@ -137,7 +137,7 @@ void FirstGpgpu::onInitialization()
 
 		// Setup
 		Renderer::RootSignatureBuilder rootSignature;
-		rootSignature.initialize(glm::countof(rootParameters), rootParameters, 0, nullptr, Renderer::RootSignatureFlags::ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT);
+		rootSignature.initialize(static_cast<uint32_t>(glm::countof(rootParameters)), rootParameters, 0, nullptr, Renderer::RootSignatureFlags::ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT);
 
 		// Create the instance
 		mRootSignature = mRenderer->createRootSignature(rootSignature);
@@ -159,7 +159,7 @@ void FirstGpgpu::onInitialization()
 			0											// instancesPerElement (uint32_t)
 		}
 	};
-	const Renderer::VertexAttributes vertexAttributes(glm::countof(vertexAttributesLayout), vertexAttributesLayout);
+	const Renderer::VertexAttributes vertexAttributes(static_cast<uint32_t>(glm::countof(vertexAttributesLayout)), vertexAttributesLayout);
 
 	{ // Create vertex array object (VAO) for content generation
 		// Create the vertex buffer object (VBO)
@@ -185,7 +185,7 @@ void FirstGpgpu::onInitialization()
 				sizeof(float) * 2	// strideInBytes (uint32_t)
 			}
 		};
-		mVertexArrayContentGeneration = mBufferManager->createVertexArray(vertexAttributes, glm::countof(vertexArrayVertexBuffers), vertexArrayVertexBuffers);
+		mVertexArrayContentGeneration = mBufferManager->createVertexArray(vertexAttributes, static_cast<uint32_t>(glm::countof(vertexArrayVertexBuffers)), vertexArrayVertexBuffers);
 	}
 
 	{ // Create vertex array object (VAO) for content processing
@@ -213,7 +213,7 @@ void FirstGpgpu::onInitialization()
 				sizeof(float) * 2	// strideInBytes (uint32_t)
 			}
 		};
-		mVertexArrayContentProcessing = mBufferManager->createVertexArray(vertexAttributes, glm::countof(vertexArrayVertexBuffers), vertexArrayVertexBuffers);
+		mVertexArrayContentProcessing = mBufferManager->createVertexArray(vertexAttributes, static_cast<uint32_t>(glm::countof(vertexArrayVertexBuffers)), vertexArrayVertexBuffers);
 	}
 
 	// Create the programs: Decide which shader language should be used (for example "GLSL" or "HLSL")
